Moving UIView & CALayer extensions to a separate source file.
This means that projects which wish to include the Neon source code directly
can avoid the "View" typealias from polluting the namespace.
